ai-sr-business-req-analyst/requirements.txt
fastapi>=0.104.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
openai>=1.3.0
uvicorn[standard]>=0.24.0
requests>=2.31.0
python-multipart>=0.0.6

